+++ client-gwt/src/main/java/org/realityforge/replicant/client/net/gwt/AbstractFrontendContextImpl.java
@@ -2,6 +2,7 @@ package org.realityforge.replicant.client.net.gwt;
 
 import javax.annotation.Nonnull;
 import org.realityforge.replicant.client.EntityRepository;
+import org.realityforge.replicant.client.EntitySubscriptionManager;
 import org.realityforge.replicant.client.runtime.AreaOfInterestService;
 import org.realityforge.replicant.client.runtime.BaseRuntimeExtension;
 import org.realityforge.replicant.client.runtime.ContextConverger;
@@ -11,17 +12,20 @@ public abstract class AbstractFrontendContextImpl
   implements BaseFrontendContext, BaseRuntimeExtension
 {
   private final EntityRepository _repository;
+  private final EntitySubscriptionManager _subscriptionManager;
   private final ReplicantClientSystem _replicantClientSystem;
   private final AreaOfInterestService _areaOfInterestService;
   private final ContextConverger _converger;
 
   public AbstractFrontendContextImpl( @Nonnull final ContextConverger converger,
                                       @Nonnull final EntityRepository repository,
+                                      @Nonnull final EntitySubscriptionManager subscriptionManager,
                                       @Nonnull final ReplicantClientSystem replicantClientSystem,
                                       @Nonnull final AreaOfInterestService areaOfInterestService )
   {
     _areaOfInterestService = areaOfInterestService;
     _repository = repository;
+    _subscriptionManager = subscriptionManager;
     _replicantClientSystem = replicantClientSystem;
 
     _converger = converger;
@@ -71,6 +75,13 @@ public abstract class AbstractFrontendContextImpl
   public EntityRepository getRepository()
   {
     return _repository;
+  }
+
+  @Nonnull
+  @Override
+  public EntitySubscriptionManager getSubscriptionManager()
+  {
+    return _subscriptionManager;
   }
 
   protected void initialSubscriptionSetup()